+++ amazon/sts/fake/src/examples/kotlin/using_provided_clients.kt
@@ -10,8 +10,8 @@ fun main() {
     val fakeSts = FakeSTS()
 
     fakeSts.client().assumeRole(
-        ARN.of(
-            Region.of("ldn-north-1"), AwsService.of("kms"), "key", "resource", AwsAccount.of("0")
-        ), UUID.randomUUID().toString()
+            ARN.of(
+                AwsService.of("kms"), Region.of("ldn-north-1"), AwsAccount.of("0"), "key", "resource"
+            ), UUID.randomUUID().toString()
     )
 }