+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v2.7.0.0 (uncut)
+- **http4k-connect-amazon-*** : Refined ARN model.
+
 ### v2.6.0.0
-- **http4k-connect-*** : API improvements for all AWS services.
+- **http4k-connect-amazon-*** : API improvements for all AWS services.
 - **http4k-connect-*** : `defaultPort()` -> `defaultPort`
 
 ### v2.5.1.0